refactor: Move global variables to root element
