+++ assets/js/main.js
@@ -10,4 +10,4 @@ import "./interactiveMap.js";
 import "./expander.js";
 import "./dialog.js";
 import "./fipValidityComparison.js";
-import "./mobileSearchButton.js";
+import "./search.js";
