+++ assets/js/search.js
