+++ assets/sass/_variables.scss
@@ -57,12 +57,39 @@ $highlight-colors-dark: (
 );
 
 :root {
-  --pf-font: roboto, Arial, Helvetica, sans-serif;
-  --pf-summary-font-size: 1.5rem;
-  --pf-result-title-font-size: 1.8rem;
-  --pf-result-excerpt-font-size: 1.5rem;
+  --font-family: "Roboto", Arial, Helvetica, sans-serif;
+
+  --link-default: #{$link-default};
+  --link-hovered: #{$link-hovered};
+  --link-special: #{$link-special};
+
+  --bg-default: #{$bg-default};
+  --bg-neutral: #{$bg-neutral};
+  --bg-accent: #{$bg-accent};
+  --bg-code: #{$bg-code};
+
+  --color-onLight: #{$color-onLight};
+  --color-table-border: #{$color-table-border};
+  --color-body: rgb(33, 37, 41);
+  --color-success: #096640;
+  --color-error: #ad1731;
+
+  --border-radius-s: 0.4rem;
+  --border-radius-m: 0.8rem;
+  --border-radius-l: 1.6rem;
+
+  --box-shadow: 0 0.4rem 1rem rgba(0, 0, 0, 0.25);
+  --box-shadow-light: 0.4rem 0.4rem 0.4rem rgba(0, 0, 0, 0.1);
+
+  --border: 0.1rem solid transparent;
+  --border-visible: 0.1rem solid #d0d0d0;
+  --outline-focus-indicator: #257fa8;
+
+  @media print {
+    --bg-neutral: white;
+  }
+
   --container-max-width: 1320px;
-  --pf-modal-max-width: var(--container-max-width);
 
   @media (max-width: #{$breakpoint-xxl}) {
     --container-max-width: 1140px;
@@ -83,21 +110,12 @@ $highlight-colors-dark: (
   @media (max-width: #{$breakpoint-sm}) {
     --container-max-width: calc(100vw - 3.2rem);
   }
-}
 
-html {
-  --link-default: #{$link-default};
-  --link-hovered: #{$link-hovered};
-  --link-special: #{$link-special};
-  --bg-default: #{$bg-default};
-  --bg-neutral: #{$bg-neutral};
-  --bg-accent: #{$bg-accent};
-  --bg-code: #{$bg-code};
-  --color-onLight: #{$color-onLight};
-  --color-table-border: #{$color-table-border};
-  --color-body: rgb(33, 37, 41);
-  --color-success: #096640;
-  --color-error: #ad1731;
+  --pf-font: var(--font-family);
+  --pf-summary-font-size: 1.5rem;
+  --pf-result-title-font-size: 1.8rem;
+  --pf-result-excerpt-font-size: 1.5rem;
+  --pf-modal-max-width: var(--container-max-width);
 
   @each $name, $color in $fip-validity-colors {
     --fip-validity-#{$name}: #{$color};
@@ -112,29 +130,18 @@ html {
     --highlight-#{$name}-bg: #{mix(white, $color, 90%)};
     --highlight-#{$name}-color: #{$color};
   }
-
-  --border-radius-s: 0.4rem;
-  --border-radius-m: 0.8rem;
-  --border-radius-l: 1.6rem;
-  --box-shadow: 0 0.4rem 1rem rgba(0, 0, 0, 0.25);
-  --box-shadow-light: 0.4rem 0.4rem 0.4rem rgba(0, 0, 0, 0.1);
-  --border: 0.1rem solid transparent;
-  --border-visible: 0.1rem solid #d0d0d0;
-  --outline-focus-indicator: #257fa8;
-
-  @media print {
-    --bg-neutral: white;
-  }
 }
 
-html[data-theme="dark"] {
+:root[data-theme="dark"] {
   @media screen {
     --link-default: #ff6b3d;
     --link-hovered: #ff8a5b;
     --link-special: #ffffff;
+
     --bg-default: #151b23;
     --bg-neutral: #0d1117;
     --bg-accent: #86761a;
+
     --color-onLight: #ffffff;
     --color-table-border: #555;
     --color-body: #e0e0e0;
