+++ assets/sass/fonts.scss
@@ -123,7 +123,7 @@
 
 body,
 button {
-  font-family: "Roboto", Arial, Helvetica, sans-serif;
+  font-family: var(--font-family);
   word-wrap: break-word;
   hyphens: auto;
 }
